remove fixme
We need the signal in this API because it allows us to cancel an in-progress fetch. Relying on `return` behavior alone does not work because it will still wait for any async tasks in the function body to finish before forcing the generator to return, and also because the FoxgloveDataPlatformPlayer calling code calls abort() outside of the for-await flow while a task may be in flight.