Add abort signal to getBackfillMessages
When in backfill mode and the user clicks to seek elsewhere, we need to abort the current request
and start a new backfill. This adds an abort controller to signal the abort for the fetch request.

This also adds an emitState prior to any backfill to give immediate feedback in the UI for the new
seek location even if the player needs to buffer some before it can play.

Fixes: #3064 #3063