Use the basename from node playground scripts in the editor
The editor uses the script filePath to resolve additional imports. When a user renames their node and removes the /studio_node/ prefix, it causes the relative path of the file to change.

To allow users to name their node whatever they want, we use only the basename of the path and manually prefix the /studio_node/ path since this is an internal implementation detail.

This change also removes the /studio_node/ prefix from new node names since it is no longer a requirement to have this prefix.

Fixes: #3148