Enable Experimental bag player as the default
The experimental bag player is ready to graduate to the default. With this player users get better
seek behavior. When a user seeks, the player will backfill the last message on every topic. This
results in a better experience for panels which rely on having the last message on topics to display
useful info (which is basically every panel).

The previous bag player would backfill messages within a certain small window prior to the seek position.
If a message was outside that window it would not get backfilled. This meant some topics with infrequent publishing
would not be backfilled and some panels would not show any data for such topics on a seek. This behavior is different
to what the panel would show had the user played the data forward from the start to the seek time.

Fixes: #2906