Fix the tick handling for IterablePlayer
When performing a _tick_ the iterable player would fail to track the correct _next_ start time for an iterator if there were no messages from the current iterator. This change fixes this logic to always set the next start time to the end of the iterator unless there are messages.

Fixes: #3285