Avoid addMessages in plot panel depending on blocks
During initial datasource loading, blocks are changing quickly. This resulted in the addMessages function constantly changing because it had logic that depended on ignoring paths that are present in blocks. useMessageReducer doesn't like that and produces a lot of warnings to the console because it keeps re-running the reducer.

This change memoizes the paths that exist in blocks and uses this variable for the addMessages logic rather than all the block data. These paths are more stable and do not change as blocks are loading.