Fix bug with node playground scripts and custom datatypes
Node playground scripts that used the ./types module and referenced custom datatypes did not work. They would fail to compile because the custom datatypes were missing from the Messages generic (i.e. Message<"my_msgs/Foo"> would be a compile error). Node playground player and transform worker logic would exclude any nodes that failed to compile when building a list of topics to subscribe to. Without subscribing to the input topics of the node script, the compilation step could never get the custom datatypes and the node would never compile successfully.

This change alters the behavior of UserNodePlayer to subscribe to the related input topics of an output topic from node scripts even if the node does not compile (but is able to identify the input topics). By subscribing to the input topics, the underlying player can provide the datatype information for the custom datatype which rebuilds the ./types module and allows the node script to compile successfully.

Prior to this change, a node which did not compile would not cause subscriptions to occur to its input topics since a node that doesn't compile cannot run. After this change, the underlying player will subscribe to "broken" nodes which may result in subscriptions to messages that are not processed by the node since the node is broken. We assume the desired state for node scripts is to be "working", we consider this an acceptable tradeoff to supporting custom datatypes.