MemoryCacheDataProvider: send notification on parse error
Previously when using the legacy MCAP player, any errors parsing
messages would be ignored from the `_provider.getMessages` result.

This PR logs these errors to console, and also sends a notification
including the error details. This uses the legacy `sendNotification`
API that we're trying to remove the use of, which is not ideal.
However:

1. This problem is fixed already with the new buffered iterable player.
2. Once we switch over to the buffered iterable player, we can happily
   delete this code and never think about it again.