Remove disabled message order feature
In https://github.com/foxglove/studio/pull/3311 we moved the playback message order setting behind a feature flag and turned it off by default. However, we didn't notice that we only made the setting available in dev mode. So no one could turn it back on in production.

Given that we've had no requests for this feature and our reasons for disabling it remain true, it is time to fully remove it.

This change removes the feature flag and all support code related to the header stamp playback feature.