Fixes placeholder handling for settings input types
Prior to this change the `placeholder` field was present for all settings input types. This change alters the type defunitions to include `placeholder` only in the input types that support it.

This change also adds placeholder array support to vec2 and vec3 input types. It also fixes a bug where vec2 and vec3 would not show the NumericInput elements if values was not specified.

Stories added to test the new behaviors.

Fixes: #3822