Catch datatype construction errors in user scripts
Rather than throwing an error when a user script cannot detect the return type, catch the error and report it as an unknown diagnostic. This allows other user scripts to run and also allows the player to run. Prior to this change, a failure in the detection would "lock up" any playback.

Fixes: #3862