Switch BufferedIterableSource to a VecQueue
shifting from an array is extremely inefficient for any non-trivially sized array because all the elements need adjusting. Since the BufferedIterableSource is constantly shifting cached messages this produces lag during playback when many messages are cached. This changes the behavior from a shift to a queue implementation which does not shift on every dequeue.