Add untilTime option to Player#startPlayback method
Our current behavior for forward seeking (via arrow key or seek button), is to "play" the data forward at least the seek amount. The purpose of this behavior is to ensure replay of stateful messages (like tf and markers) so pressing "seek" from the start of the file results in the same playback as playing from the start of the file.

Unfortunately, the way we implemented this feature was to start playback and attempt to pause playback after the seek target was reached. This behavior would result in a forward seek moving forward different and variable amounts of time.

This change removes this workaround play/quick-pause behavior in favor of first-class support within the player. The player is startPlayback method is passed a new option argument with an optional `untilTime`. This option tells the player that it should start playback but stop when the `untilTime` is reached.

A player is free to implement this however it wants. This change updates the RandomAccessPlayer to use a seek and the IterablePlayer to play forward untilTime is reached.

Fixes: #3980
Fixes: #3872