Wait for previous emit state in tick
In the iterable player, when the previous message is still past the target tick time, the logic would override the messages and current time without waiting on the current emitState to finish. This would result in overriding outgoing messages.

This change fixes this behavior to wait on any active emit state to finish before proceeding to emit state for the new message data.

Fixes: #4073
Fixes: #4057