Fix CachingIterableSource getBackfillMessages
Similar to https://github.com/foxglove/studio/pull/4126, the CachingIterableSource would skip over some exactly matching timestamps when looking at the cache block for finding backfill messages. This meant the cache might have a message we could use for backfill but we would miss using it which results in worse cache utilization. It can also result in loading a message that is older than the last message.