Fix manual looping when using the latching player.
PR https://github.com/foxglove/studio/pull/4045 removed the resumePlay function which was responsible for seeking to the start and playing when the user wanted to continue playing from the end. This change adds that functionality back so when a user wants to "play" and playback is at the end, we seek to the start.

Unfortunately, this logic was being intercepted by logic in RepeatAdapter which would immediately try to pause playback because repeat was off. This change updates the IterablePlayer to indicate it is no longer playing when it has reached the end and removes the RepeatAdapter in favor of a message pipeline selector that correctly handles seeking to the start if repeat is enabled.

Fixes: #3921