Improve error message when user script errors during runtime
When a user script errors during runtime and fails to produce a message, the previous error displayed would only indicate that a script failed to produce a message. This warning did not make it obvious that the user script in fact errored during runtime.

This change improves the error to be clear that the script errored during runtime when there is an error rather than only a missing message.

Fixes: #4084