Call pauseFrame/resumeFrame in PanelExtensionAdapter
When handling context updates it is important to call pauseFrame to indicate that the panel will perform some work to render the frame and resumeFrame when the panel's work is complete. Without these calls, the message pipeline will assume the panel has completed the frame and will feed another frame to the panel to render. This additional frame will override the previous and result in dropped messages for the panel.

Fixes: #4091