[New3D] Create coordinate frames from all <link>s when loading URDFs
Previously, we would only recognize new coordinate frames in URDFs from joints (which define a parent/child relationship between two links). If a URDF defined a single <link> with no <joint>, or a link that was disconnected from any joint, it would not end up defining a coordinate frame in Studio and the URDF could not be rendered.