Add "todo" to list of disallowed words in comments
Similar to "fixme", this change adds "todo" to the list of disallowed words in comments.

Rather than use TODO comments in the code, we create tickets for any follow-up work that needs to happen. Tickets are how we review outstanding work and ensure it is prioritized and completed.

If there is an implementation tradeoff, a code comment should describe the comment and and references (which could be tickets too) but it should not do this in the form of a TODO.