Fix extension panel config reset bug on player change
There were two layered issues leading to a reset of extension panel config when changing data sources.

The first was a change in Workspace to remount the panels when setSubscriptions from the message pipeline changes. This change was made in https://github.com/foxglove/studio/pull/4292 when requestBackfill was removed. Unfortunately, setSubscriptions is stable and does not change when the player changes. This meant we were not actually getting a new value to remount. The fix is to switch to using playerId which does change when the player changes.

The second bug occurs inside of the panel extension adapter. Due to some react update ordering, the panelExtensionContext was being re-created when the player changed. During this rebuild, the initialState would be provided to the new panelExtensionContext. However this initialState was the original and empty state for the panel because we needed a stable variable for the initialState. The fix is to make initialState a ref so that whenever the panel does initialize it does so with the latest configuration value.