Fix playback when changing topic subscriptions
PR https://github.com/foxglove/studio/pull/4379 broke subscribing to topics when paused. This behavior broke because 4379 added a check to seekPlayback that would avoid seeking if the requested seek time was already the current time. Unfortunately, the IterablePlayer relied on this behavior when subscriptions changed to trigger a backfill and reset the message iterators by calling seekPlayback with the currentTime.

This change fixes the setSubscriptions behavior to instead directly request the seek-backfill state rather than routing through seekPlayback which is a public-facing api.

Fixes: #4395