Add geojson message to correct feature group
The previous behavior of adding a geojson message would always add the message to the allFrames feature group. When connected to a live data source this would result in the geojson messages accumulating since allFrames would not be cleared when receiving new messages.

This change updates the logic for adding geojson messages to ensure they are added to the correct feature group (allFrames or currentFrame). Each feature group is cleared when processing messages for that specific feature group thus avoiding stale messages in the feature group.

Fixes: #4360