Fix block size tracking for BlockLoader
Fix incorrect sizeInBytes tracking for BlockLoader when closing out the current block or filling block gaps. Prior to this change the sizeInBytes (which tracks the size all messages within a block), would incorrectly grow when finishing a block and moving to the next block. This would result in blocks having a sizeInBytes that was a sum of all previous block sizes. When calling cacheSize this would result in a large cache size for totalSizeInBytes which would attemt to evict all blocks.

This change also updates streamMessages to remove the abort signal event listener when exiting normally to avoid printing a misleading "Manual abort of streamMessages" log.

Fixes: #4466