Make a copy of each rosbag message event data
Each rosbag message event contains the raw data for the message. This data is a view on top of the chunk and keeps a reference to the entire chunk data for the lifetime of the message data.

For Studio users are often displaying only a subset of the messages within a file. As we load messages we cache them in memory. When every message keeps a reference to the original chunk this results in large memory use even when viewing one or two topics. For large bag files this memory use can grow so large that eventually causes Studio to crash.

Even tho Studio playback does have a caching layer with a limit of memory use, this caching layer looks at the individual message data size. This does not take into account any underlying shared buffers or references for that data.