Add `legacyIds?` to IDataSourceFactory
When we combined the Ros1Remote and McapRemote factories into a single factory we lost their respective factory `id` fields. These fields are used in deep-link urls to specify which factory handles the rest of the query args. This change broke existing URLs using these ids.

This fixes the broken behavior by introducing a new optional `legacyIds` field to IDataSourceFactory and updating the selectSource(id) function in PlayerManager to lookup into these ids when selecting a source. Users are able to use the old ids and the new id during the transition period.

Fixes: #4937