Move GlobalCss back into <App> under a prop
The MuiModal root appears outside of the #root component directly within body. This means it does not inherit the styling we apply on #root (such as font family and sizes) for some components (Link seems to be one of them).

As a workaround we move <GlobalCss> back into <App> under a property to allow users of <App> to specify if they want to turn these global values on. Without these the styling might be wrong on certain elements if the user does not set the same font on the body that they use within the app.

Additionally, GlobalCss must exist within the ThemeProvider and was not before so its setting of the background color and body elements was incorrect when used outside of <App>

Fixes: #4934