Always emit urlState in FoxgloveWebsocketPlayer
FoxgloveWebsocketPlayer was not emitting urlState when it did not have a connection. This prevented downstream logic from updating the URL with the latest player information. This chagne updates the player to always emit urlState which causes the app to update the URL bar immediately once the data source is selected.