Fix image rendering in 3d panel
This change resolves a regression introduces in https://github.com/foxglove/studio/pull/5555. The change in that PR incorrectly skipped updating the image when the _recomputeCameraModel returned undefined - thinking that it would always return the latest camera model.

This fix modes the _recomputeCameraModel logic into _updateImageRenderable and changes the inputs to _updateImageRenderable to accept the CameraInfo rather than the camera model. It also updates _recomputeCameraModel to return the cached camera model if available.

_recomputeCameraModel was originally designed to return undefined so handleCameraInfo could avoid rendering the image for camera calibration messages which did not result in changes to the camera model. That logic changed in #5555 and the handleCameraInfo handler no longer renders images but instead saves the latest camera info and leaves image rendering to the handleImage handler.