Rework the TimeBasedChart downsample code organization
Rather than having two layers of throttle and debounce, this change moves uses one layer of throttle and a Downsampler class to manage the state for the downsample. useEffects update the Downsampler instance with the latest state information and the debounced callback triggers a downsample which updates the state variable with the downsampled dataset.