Fix memory leak with accumulated condvar callbacks in UserNodePlayer
UserNodePlayer message handling uses a Condvar to signal when the node should terminate (maybe due to code changes in other nodes or itself). The message processing logic calls `.wait()` on this condvar for every message to process and races the returned signal against the RPC call to process. During normal runtime (with no changes to user scripts), nothing notifies all of these signals and they continue to pile up in the callback queue of the condvar.

This change updates the logic to create one signal during the node registration for use by the message processor callback. This signal is re-created on terminate.