Mark `subscribe(topics[])` as deprecated.
Subscribing to an array of topics does not allow for controlling whether you want partial or full subscriptions and instead assumes full subscriptions which is often more than you want to subscribe to. By deprecating this form of subscribe we lead users to the more flexible form with subscriber options.