Add "resetDeepLinks" to Desktop interface
The resetDeepLinks function resets the deep links for the desktop app by using a cookie + reload hack. Since deep links are only read on startup
we need to avoid reading them if we want to ignore them in App.

This change also extracts the root element of the AppBar into its own component so we can use it without using the entire AppBar. We need this since our desktop app disables system titlebar and we still want an AppBar background when we want to avoid rendering the workspace which renders the AppBar.