Skip undefined messages from message converters
This change updates the message converter processing logic to ignore _undefined_ results. Prior to this
change when a message converter returned _undefined_ the _undefined_ value would be considered the message and
provided to the panel in the currentFrame and allFrames.

This change updates the processing logic to detect _undefined_ results from message converters and skip adding
those to the currentFrame and allFrames for panels.